+++ SupportYoutubeContentCreators.js
@@ -1,15 +1,15 @@
 // ==UserScript==
 // @name         Support Youtube Content Creators
 // @author       antipatico 
-// @version      0.2.2-beta
+// @version      0.3-beta
 // @description  Whitelisting your loved content creators videos from your ADBlocker is now as easy as a click! Tested with Firefox + uBlock Origin + Violentmonkey.
 // @homepageURL  https://github.com/antipatico/support-youtube-content-creators
 // @updateURL    https://raw.githubusercontent.com/antipatico/support-youtube-content-creators/master/SupportYoutubeContentCreators.js
 // @downloadURL  https://raw.githubusercontent.com/antipatico/support-youtube-content-creators/master/SupportYoutubeContentCreators.js
 // @grant        GM_getValue
 // @grant        GM_setValue
+// @grant        GM_registerMenuCommand
 // @include      /^https?://(www\.)?youtube\.com/.*$/
-// @require      https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js
 // @run-at       document-idle
 // @noframes
 // ==/UserScript==
@@ -55,221 +55,228 @@ Take in consideration the possibility to offer me a beer.
 
 "use strict";
 
-/* Enums */
-let LogLevel = { error: 0, warning: 1, info: 2, debug: 3};
+/* Log module */
+let log = {
+  /* Log levels constants */
+  get ERROR() { return 0; },
+  get WARNING() { return 1; },
+  get INFO() { return 2; },
+  get DEBUG() { return 3; },
+  
+  get verbosity() {
+    return this.INFO; // Change log verbosity from here.
+  },
+  
+  /* Log functions, basic functions */
+  log: function() {
+    var args = [].slice.call(arguments);
+    args.unshift('%c[SYCC]:', 'font-weight: bold;color: #ef9f09;');
+    console.log.apply(console, args);
+  },
+  error: function(message) { this.log("Error: "+message); },
+  warning: function(message) { if(this.verbosity >= this.WARNING) this.log("Warning: "+message); },
+  info: function(message) { if(this.verbosity >= this.INFO) this.log(message); },
+  debug: function(message) { if(this.verbosity >= this.DEBUG) this.log(message); } 
+}
 
-/* Constants */
-let DEFAULT_TAG = "support";
-let LOG_LEVEL = LogLevel.info; // change this to debug if you want a verbose output.
-/* Regex used accross the script */
-let REGEX_TAG = /^\w+$/;
-let REGEX_CHANNEL = /\/channel\/([A-Za-z0-9\-_]+)/;
-let REGEX_USER = /\/user\/(\w+)/;
-let REGEX_USER_FEATURED = /\/user\/(\w+)\/featured/;
-let REGEX_VIDEO = /\/watch\?v=\w+/;
-/* UI Constants */
-let TEXT_LOVED = "🧡";
-let TEXT_HATED = "🖤";
-let TIP_SUPPORT = "Click to whitelist ADS!";
-let TIP_UNSUPPORT = "Click to blacklist ADS!";
-let BTN_VIDEO_TEMPLATE = "<div style='cursor:pointer'></div>";
-let BTN_CHANNEL_TEMPLATE = "<span style='cursor:pointer' class='style-scope ytd-c4-tabbed-header-renderer'></span>";
-let BTN_VIDEO_ID = "sycc-video-btn";
-let BTN_CHANNEL_ID = "sycc-channel-btn";
 
-/* Log functions */
-function log(message) { console.log("[SYCC] " + message); }
-function error(message) { log("Error: "+message); }
-function warning(message) { if(LOG_LEVEL >= LogLevel.warning) log("Warning: "+message); }
-function info(message) { if(LOG_LEVEL >= LogLevel.info) log(message); }
-function debug(message) { if(LOG_LEVEL >= LogLevel.debug) log("Debug: "+message); }
-
-/* Object representing persistent data */
+/* Persistent data module */
 let data = {
+  /* Constant fallback tag */
+  get DEFAULT_TAG() { return "support"; },
+  
+  /* Used to sanitize custom tags */
+  get REGEX_TAG() { return /^\w+$/; },
+  
+  /* Sanitize the custom tag (custom tag not yet implemented) */
+  sanitizeTag: function(tag) {
+    return (this.REGEX_TAG.test(tag))?tag:this.DEFAULT_TAG;
+  },
+
   get tag() {
-    return sanitizeTag(GM_getValue("tag", DEFAULT_TAG));
+    return this.sanitizeTag(GM_getValue("tag", this.DEFAULT_TAG));
+  },
+  
+  set tag(val) {
+    GM_setValue("tag", this.sanitizeTag(val));
   },
   
   get htag() {
     return "#" + this.tag;
   },
   
-  get whitelist() {
-    return GM_getValue("whitelist", []);
+  whitelist: {
+    get: function() { return GM_getValue("whitelist", []); },
+    
+    set: function(val) {
+      if(Array.isArray(val)) {
+        GM_setValue("whitelist", val);
+      } else {
+        log.error("Passed value to set whitelist(val) is NOT an array");
+      }
+    },
+    
+    contains: function(cc) {
+      return data.whitelist.get().includes(cc);
+    },
+    
+    append: function(cc) {
+      if(!data.whitelist.contains(cc)) {
+        let wl = data.whitelist.get();
+        wl.push(cc);
+        data.whitelist.set(wl);
+      }
+    },
+    
+    remove: function(cc) {
+      if(data.whitelist.contains(cc)) {
+        let wl = data.whitelist.get();
+        wl.pop(cc);
+        data.whitelist.set(wl);
+      }
+    },
+    
+    clear: function() {
+      data.whitelist.set([]);
+    }
+  }
+}
+
+/* Core module */
+let core = {
+  /* Constants used across the script */
+  get REGEX_CHANNEL() { return /\/channel\/([A-Za-z0-9\-_]+)/; },
+  get REGEX_USER() { return /\/user\/(\w+)/; },
+  get REGEX_USER_FEATURED() { return /\/user\/(\w+)\/featured/; },
+  get REGEX_VIDEO() { return /\/watch\?v=\w+/; },
+  
+  /* Location identifying functions */
+  inVideoPage: function() { return this.REGEX_VIDEO.test(document.location); },
+  inChannelPage: function() { return this.REGEX_CHANNEL.test(document.location); },
+  inUserPage: function() { return this.REGEX_USER.test(document.location); },
+  inFeaturedPage: function() { return this.REGEX_USER_FEATURED.test(document.location); },
+
+  /* Retrieve the current content creator ID from the current page
+   * Used widely accross the script to retrieve the current content creator.
+   * 
+   * Works on:
+   * + Channel pages (/channel/ID)
+   * + User pages (/user/USERNAME)
+   * + Video pages (/watch?v=VIDEOID)
+   * 
+   * If not in these pages or not found, return NULL
+   */
+   getContentCreator: function() {
+    log.debug("Attempting to retrieve content creator for this page");
+    let channelURL = "";
+
+    if (core.inChannelPage()) {
+      log.debug("Detected channel page");
+      channelURL = document.location.href;
+    } else if (core.inUserPage()) {
+      channelURL = document.querySelector("meta[property='og:url']").content;
+    } else if (core.inVideoPage()) {
+      let DOMel = document.querySelector("yt-formatted-string#owner-name").firstChild;
+      if(DOMel == undefined) {
+        log.error("Can't find channel URL inside video page"); // with new hooks it should never happen.
+        return null;
+      }
+      channelURL = DOMel.href;
+    } else {
+      log.debug("Not in a content page, no REGEX matched");
+    }
+    return (channelURL)?channelURL.match(this.REGEX_CHANNEL)[1]:null;
+  },
+
+  /* Function responsible to redirect in case of whitelisted content */
+  redirectStep: function() {
+     /* 1. Check if we are whitelisted already
+      * 2. Retrieve content creator
+      * 2. Is the content creator whitelisted?
+      * 3. If yes redirect to tagged URL
+      */
+    if (document.location.href.includes(data.htag)) {
+      log.info("Already in whitelist, nothing to do :)");
+      return;
+    }
+
+    /* Disable redirecting in non-content page */
+    if(core.inUserPage() && !core.inFeaturedPage()) {
+      return;
+    }
+
+    log.debug("URL not whitelisted (yet), starting redirectStep!");
+    let cc = core.getContentCreator();
+    log.debug("Content creator: "+cc);
+    if(cc && data.whitelist.contains(cc)) {
+      log.info("Content creator in whitelist, redirecting..");
+      document.location.assign(document.location.href + data.htag);
+    } else  {
+      if(cc) {
+        log.info("Content creator not in whitelist");
+      }
+    }
+  },
+
+  /* Add or remove a content creator from the whitelist */
+  toggleWhitelist: function(cc) {
+    //let cc = event.data.cc;
+    let refreshURL = "";
+    if(cc && data.whitelist.contains(cc)) {
+      log.debug("Removing content creator to whitelist");
+      data.whitelist.remove(cc);
+      refreshURL = document.location.href.replace(data.htag, "");
+    } else {
+      log.debug("Adding content creator from whitelist");
+      data.whitelist.append(cc);
+      refreshURL = document.location.href + data.htag;
+    }
+    document.location.assign(refreshURL);
+  }
+}
+
+/* Commands module */
+let commands = {
+  init: function() {
+    GM_registerMenuCommand("[SYCC] Change whitelisted tag", commands.onChangeTag);
+    GM_registerMenuCommand("[SYCC] Toggle whitelist for current Content Creator (will refresh the page)", commands.onToggleWhitelist);
+    GM_registerMenuCommand("[SYCC] Clear whitelist", commands.onClearWhitelist);
   },
   
-  set whitelist(val) {
-    if(Array.isArray(val)) {
-      GM_setValue("whitelist", val);
-    } else {
-      error("Passed value to set whitelist(val) is NOT an array");
+  onChangeTag: function() {
+    let input = prompt('The current tag is "' + data.tag + '".\n Enter the new desired tag');
+    if(input && input != "") {
+      data.tag = input;
+    }
+  },
+  
+  onToggleWhitelist: function() {
+    let cc = core.getContentCreator();
+    if(!cc) {
+      alert("Can't understand who you want to whitelist. Try again in a video page!");
+      log.warning("Content Creator not found!");
+      return;
+    }
+    core.toggleWhitelist(cc);
+    log.info(data.whitelist.contains(cc)?"Successfully added to the whitelist":"Content Creator removed from the whitelist"); 
+  },
+  
+  onClearWhitelist: function() {
+    if(confirm("Are you sure you want to empty your whitelist?\nYou won't be able to recover your current whitelist anymore!")) {
+      data.whitelist.clear();
     }
   }
-}
-
-/* Location identifying functions */
-function inVideoPage() { return REGEX_VIDEO.test(document.location); }
-function inChannelPage() { return REGEX_CHANNEL.test(document.location); }
-function inUserPage() { return REGEX_USER.test(document.location); }
-function inFeaturedPage() { return REGEX_USER_FEATURED.test(document.location); }
-
-/* Sanitize the custom tag (custom tag not yet implemented) */
-function sanitizeTag(tag) {
-  return (REGEX_TAG.test(tag))?tag:DEFAULT_TAG;
-}
-
-/* Retrieve the current content creator ID from the current page
- * Used widely accross the script to retrieve the current content creator.
- * 
- * Works on:
- * + Channel pages (/channel/ID)
- * + User pages (/user/USERNAME)
- * + Video pages (/watch?v=VIDEOID)
- * 
- * If not in these pages or not found, return NULL
- */
-function getContentCreator() {
-  debug("Attempting to retrieve content creator for this page");
-  let channelURL = "";
-  
-  if (inChannelPage()) {
-    debug("Detected channel page");
-    channelURL = document.location.href;
-  } else if (inUserPage()) {
-    channelURL = $("meta[property='og:url']").attr("content");
-  } else if (inVideoPage()) {
-    let DOMel = $("div#owner-container").children("yt-formatted-string#owner-name");
-    if(DOMel == undefined) {
-      error("Can't find channel URL inside video page"); // with new hooks it should never happen.
-      return null;
-    }
-    channelURL = DOMel.children().attr("href");
-  } else {
-    debug("Not in a content page, no REGEX matched");
-  }
-  
-  return (channelURL)?channelURL.match(REGEX_CHANNEL)[1]:null;
-}
-
-/* Function responsible to redirect in case of whitelisted content */
-function redirectStep() {
-   /* 1. Check if we are whitelisted already
-    * 2. Retrieve content creator
-    * 2. Is the content creator whitelisted?
-    * 3. If yes redirect to tagged URL
-    */
-  if (document.location.href.includes(data.htag)) {
-    debug("Already whitelisted, aborting redirectStep");
-    return;
-  }
-  
-  /* Disable redirecting in non-content page */
-  if(inUserPage() && !inFeaturedPage()) {
-    return;
-  }
-  
-  debug("URL not whitelisted (yet), starting redirectStep!");
-  let cc = getContentCreator();
-  debug("Content creator: "+cc);
-  if(cc && data.whitelist.includes(cc)) {
-    debug("Content creator in whitelist, redirecting..");
-    document.location.assign(document.location.href + data.htag);
-  } else  {
-    debug("Content creator not in whitelist (or not in a content page)");
-  }
-}
-
-function createButton(template, id, cc) {
-  let button = $(template);
-  let text = (data.whitelist.includes(cc))?TEXT_LOVED:TEXT_HATED;
-  let tooltip = (data.whitelist.includes(cc))?TIP_UNSUPPORT:TIP_SUPPORT;
-  button.attr("id", id);
-  button.text(text);
-  button.attr("title", tooltip);
-  button.click({cc: cc}, toggleWhitelist);
-  
-  return button;
-}
-
-/* Add or remove a content creator from the whitelist */
-function toggleWhitelist(event) {
-  let cc = event.data.cc;
-  let refreshURL = "";
-  let whitelist = data.whitelist;
-  if(cc && whitelist.includes(cc)) {
-    debug("Removing content creator to whitelist");
-    whitelist.pop(cc);
-    refreshURL = document.location.href.replace(data.htag, "");
-  } else {
-    debug("Adding content creator from whitelist");
-    whitelist.push(cc);
-    refreshURL = document.location.href + data.htag;
-  }
-  data.whitelist = whitelist;
-  document.location.assign(refreshURL);
-}
-
-/* Inject the DOM elements inside the page */
-function injectElements() {
-  /* Video page button */
-  if(inVideoPage() && !$("#"+BTN_VIDEO_ID).length) {
-    let cc = getContentCreator();
-    let button = createButton(BTN_VIDEO_TEMPLATE, BTN_VIDEO_ID, cc);
-    $("div#owner-container").append(button);
-  }
-  /* Channel / user page button */
-  else if ((inChannelPage() || inUserPage()) && !$("#"+BTN_CHANNEL_ID).length) {
-    let cc = getContentCreator();
-    let button = createButton(BTN_CHANNEL_TEMPLATE, BTN_CHANNEL_ID, cc);
-    $("h1#channel-title-container").append(button);
-  }
-}
-
-/* Update the DOM elements after navigation */
-function updateElements() {
-  let videoButton = $("#"+BTN_VIDEO_ID);
-  let channelButton = $("#"+BTN_CHANNEL_ID);
-  let cc = getContentCreator();
-  
-  if(!cc) {
-    return; // not in a content page
-  }
-  
-  let text = (data.whitelist.includes(cc))?TEXT_LOVED:TEXT_HATED;
-  let tooltip = (data.whitelist.includes(cc))?TIP_UNSUPPORT:TIP_SUPPORT;
-  
-  [videoButton, channelButton].forEach(function(button) {
-    if(button.length) {
-      button.text(text);
-      button.attr("title", tooltip);
-      button.unbind("click");
-      button.click({cc: cc}, toggleWhitelist);
-    }
-  });
-}
-
-/* Trigger a single event instead of three different ones */
-function onNavigateFinish() {
-  redirectStep();
-  injectElements();
-  updateElements();
 }
 
 /* Main function
  * First we hook the redirectStep function (used to actually whitelist stuff)
- * then we hook the injectElements function (used to create the elements to interact with the script itself)
- * last but not least we hook the updateElements (used on page refresh to edit the elements data)
+ * then we init the commands module.
  */
-function main() {
-  info("Support Youtube Content Creators by antipatico loaded, now running");
-  info("Consider to star me on GitHub!");
-  debug("jQuery version " + jQuery.fn.jquery);
-  /* Hook to first page loaded */
-  window.addEventListener("yt-service-request-completed", redirectStep);
-  window.addEventListener("yt-service-request-completed", injectElements);
-  /* Hook to new video link loaded, wait to let the youtube's scripts do their job */
-  window.addEventListener("yt-navigate-finish", function() { setTimeout(onNavigateFinish, 1); });
-}
-
-main(); // Load the payload :)
+log.info("Support Youtube Content Creators by antipatico loaded, now running");
+log.info("Consider to star me on GitHub!");
+/* Hook to first page loaded */
+window.addEventListener("yt-service-request-completed", core.redirectStep);
+/* Hook to new video link loaded, wait to let the youtube's scripts do their job */
+window.addEventListener("yt-navigate-finish", function() { setTimeout(core.redirectStep, 1); });
+commands.init();